Updated build scripts

+++ unity_rtm_sdk/Projects/Windows/build.sh
@@ -1,23 +1,38 @@
 #!/bin/bash
-## build script for RTM plugin for Android on Unity
+## ==============================================================================
+## build script for RTM plugin for Windows on Unity
+## this build script doesn't support building on Mac environment.
+## Please use the python build script and bat files on Windows PC to build the target.
+## required environmental variables:
+##   $RTM_VERSION
+## ==============================================================================
 
+PLATFORM="Windows"
 
 function download_library {
+    ARCH=$1
+    DOWNLOAD_VERSION=$2
     DOWNLOAD_URL="https://download.agora.io/rtmsdk/release"
-    x86_FILE="Agora_RTM_SDK_for_Windows_x86_Unity_v1.4.2.zip"
-    x64_FILE="Agora_RTM_SDK_for_Windows_x64_Unity_v1.4.2.zip"
+    ZIP_FILE="Agora_RTM_SDK_for_Windows_${ARCH}_Unity_${DOWNLOAD_VERSION}.zip"
 
-    if [[ ! -e $x86_FILE ]]; then
-        wget $DOWNLOAD_URL/$x86_FILE
+    if [[ ! -e $ZIP_FILE ]]; then
+        wget $DOWNLOAD_URL/$ZIP_FILE
+	status=$?
+	if [ ! $status == 0 ]; then
+	    echo "Status of wget:$status"
+	    # change from v1_4_2 to v1.4.2
+	    VERSION_FORMAT2="${DOWNLOAD_VERSION//_/.}"
+	    if [ $DOWNLOAD_VERSION == $VERSION_FORMAT2 ]; then
+		# we've tried the two formats!
+		echo "$ZIP_FILE Fail!" && exit 1
+	    else 
+		download_library $ARCH $VERSION_FORMAT2
+	    fi
+	fi
     fi
-    if [[ ! -e $x64_FILE ]]; then
-        wget $DOWNLOAD_URL/$x64_FILE
-    fi
+
     #unzip
-    rm -rf sdk agoraRTMCWrapper/sdk
-    mkdir sdk
-    (cd sdk && unzip -o ../$x86_FILE product_sdk/sdk/* && mv product_sdk/sdk x86)
-    (cd sdk && unzip -o ../$x64_FILE product_sdk/sdk/* && mv product_sdk/sdk x64)
+    (cd sdk && unzip -o ../$ZIP_FILE product_sdk/sdk/* && mv product_sdk/sdk $ARCH)
     rm -rf sdk/product_sdk
 }
 
@@ -28,10 +43,34 @@ function package_windows_project {
     zip -r $1 agoraRTMCWrapper
 }
 
+function Clean {
+    echo "removing $PLATFORM build intermitten files..."
+    rm -rf sdk *.zip Agora_RTM_SDK_for_${PLATFORM}_* 
+}
+
+if [ "$1" == "clean" ]; then
+    Clean
+    exit 0
+fi
+
+# We will require the setting of RTM_VERSION environmental variable
+if [ -z ${RTM_VERSION+x} ]; then
+    echo "ERROR, environment variable RTM_VERSION (e.g. 'v1_4_2') must be set!"
+    exit 1
+    else echo "$PLATFORM RTM_VERSION = $RTM_VERSION"
+fi
+
+DOWNLOAD_VERSION=$RTM_VERSION
+
 OUTPUT_PACKAGE="RTM_windows.zip"
 
 rm -f $OUTPUT_PACKAGE
-download_library
+rm -rf sdk agoraRTMCWrapper/sdk
+mkdir sdk
+
+download_library "x86" $DOWNLOAD_VERSION
+download_library "x64" $DOWNLOAD_VERSION
+
 package_windows_project $OUTPUT_PACKAGE
 
 echo
